requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
bcrypt==4.0.1
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
jinja2==3.1.2
requests==2.31.0
